+++ src/core/index.ts
@@ -1,4 +1,5 @@
 export * from "./app-error.js";
+export * from "./prompt-prefix.js";
 export * from "./ripgrep.js";
 export * from "./serialization.js";
 export * from "./web-search-config.js";